Did some new lessons
I'm trying to keep going with four files: A txt that keep track of the
video, a gnplot that executes whatever the video does, and a
Resumo.gnplot to organize and summarize everything, and keep it as a
prototype to a real plot script.
Everything is held together by vim_sessions. So users of VIM can open it
as a "project folder"