Licao05 and Licao06 updated
Licao6 is not my best work, but honestly, I don't care that much about
colorschemes, just the basic is enough to me, the gradient of colors is
way too much for me to handle. And this should be a preliminary lesson,
where I get used to this new tool, so...
It's just too much for now to pay attention and learn. Maybe in the
future...